Created a concrete Creep class
Group work with Johan

+++ ProjectFiles/core/src/edu/chl/proximity/Models/Creeps/Creep.java
@@ -19,7 +19,7 @@ public abstract class Creep extends BoardObject {
     private double distanceToNextWayPoint;
     private Path path;
     private Sound devolveSound;
-    private double speed;
+    private int speed;
 
     public Creep(Image image, int speed) {
